Consolidate audio interfaces: Replace Bartender Remote with Audio Control Center
- Replaced /remote page content with Audio Control Center implementation
- Removed duplicate /audio-control route (now consolidated at /remote)
- Updated navigation labels from 'Remote Control' to 'Audio Control Center'
- Updated CEC control page link to point to new consolidated interface
- Removed duplicate Music & Audio section from homepage

Benefits:
- Single unified audio control interface to maintain
- Uses dynamic Atlas configurations
- Includes Matrix routing functionality
- Cleaner interface without bottom toolbar
- All existing /remote links continue to work
